+++ Sources/TuistGraph/Models/RunAction.swift
@@ -8,6 +8,7 @@ public struct RunAction: Equatable {
     public let executable: TargetReference?
     public let filePath: AbsolutePath?
     public let arguments: Arguments?
+    public let options: RunActionOptions
     public let diagnosticsOptions: Set<SchemeDiagnosticsOption>
 
     // MARK: - Init
@@ -16,12 +17,14 @@ public struct RunAction: Equatable {
                 executable: TargetReference?,
                 filePath: AbsolutePath?,
                 arguments: Arguments?,
+                options: RunActionOptions = .init(),
                 diagnosticsOptions: Set<SchemeDiagnosticsOption>)
     {
         self.configurationName = configurationName
         self.executable = executable
         self.filePath = filePath
         self.arguments = arguments
+        self.options = options
         self.diagnosticsOptions = diagnosticsOptions
     }
 }